foi realizado: o situação no cadastro de pessoas o situação na edição de pessoas o situação no deletar pessoas o situação no consulta de pessoas o alterar situação na tela pessoas RQ01: CA.001 - OK CA.002 - OK CA.003 - resolver = botao de criar nao funciona CA.004 - resolver = botao de confirmar edição nao funciona CA.005 - OK
RQ02:
CA.001 - codigo criado=nao deu pra testar por conta do erro ao criar
CA.002 - codigo criado=nao deu pra testar por conta do erro ao editar
CA.003 - codigo criado=nao deu pra testar por conta do erro ao criar e editar
CA.004 - codigo criado=nao deu pra testar por conta do erro ao criar e editar
CA.005 - codigo criado=nao deu pra testar por conta do erro ao criar e editar
CA.006 - codigo criado=nao deu pra testar por conta do erro ao criar
CA.007 - codigo criado=nao deu pra testar por conta do erro ao criar

falta resolver:
descobrir e consertar erro ao tentar criar/editar/deletar pessoas

+++ AplicativoWebAcademia/Controllers/PessoaModelsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using AplicativoWebAcademia.Data;
 using AplicativoWebAcademia.Models;
+using AplicativoWebAcademia.regras;
 
 namespace AplicativoWebAcademia.Controllers
 {
@@ -56,12 +57,11 @@ namespace AplicativoWebAcademia.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Codigo,Nome,QuantidadeFilhos,Email,Salario,DataNascimento")] PessoaModel pessoaModel)
         {
+            pessoaModel.Alterar = "Ativo";
+            _context.Update(pessoaModel);
             return View(pessoaModel);
         }
-        public string Ativar(PessoaModel pessoaModel)
-        {
-            return (pessoaModel.Alterar.Equals("Ativo")) ? pessoaModel.Alterar = "inativo" : pessoaModel.Alterar = "Ativo";
-        }
+       
         // GET: PessoaModels/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -83,7 +83,7 @@ namespace AplicativoWebAcademia.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Codigo,Nome,QuantidadeFilhos,Email,Salario,DataNascimento")] PessoaModel pessoaModel)
+        public async Task<IActionResult> Edit(int id, [Bind("Codigo,Nome,QuantidadeFilhos,Email,Salario,DataNascimento,Alterar")] PessoaModel pessoaModel)
         {
             if (id != pessoaModel.Codigo)
             {
@@ -92,6 +92,37 @@ namespace AplicativoWebAcademia.Controllers
 
             if (ModelState.IsValid)
             {
+                if (PessoaRegras.VerificaFilhos(pessoaModel.QuantidadeFilhos))
+                {
+                    ModelState.AddModelError("regras", "A quantidade minima de filhos é zero!");
+                    return View(pessoaModel);
+                }
+                if (PessoaRegras.VerificaNascimento(pessoaModel.DataNascimento))
+                {
+                    ModelState.AddModelError("regras", "A data de nascimento deve ser igual ou superior a 01/01/1990");
+                    return View(pessoaModel);
+                }
+                if (PessoaRegras.VerificaSalarioMin(pessoaModel.Salario))
+                {
+                    ModelState.AddModelError("regras", "O salário não pode ser inferior a R$ 1.200");
+                    return View(pessoaModel);
+                }
+                if (PessoaRegras.VerificaSalarioMax(pessoaModel.Salario))
+                {
+                    ModelState.AddModelError("regras", "O salário não pode ser superior a R$ 13.000");
+                    return View(pessoaModel);
+                }
+                var pessoaEmail = _context.PessoaModel.Where(x => x.Email.Equals(pessoaModel.Email) && x.Codigo != pessoaModel.Codigo);
+                if (pessoaEmail.Count() > 0)
+                {
+                    ModelState.AddModelError("regras", "O Email já está cadastrado!");
+                    return View(pessoaModel);
+                }
+                if (PessoaRegras.VerificaInativo(pessoaModel.Alterar))
+                {
+                    ModelState.AddModelError("regras", "Não é possivel editar uma pessoa na situação 'Inativo'");
+                    return View(pessoaModel);
+                }
                 try
                 {
                     _context.Update(pessoaModel);
@@ -112,7 +143,21 @@ namespace AplicativoWebAcademia.Controllers
             }
             return View(pessoaModel);
         }
+        [HttpGet]
+        public async Task<IActionResult> Alterar(string? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
 
+            var pessoaModel = await _context.PessoaModel.FindAsync(id);
+            if (pessoaModel == null)
+            {
+                return NotFound();
+            }
+            return View(pessoaModel);
+        }
         // GET: PessoaModels/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
@@ -120,7 +165,6 @@ namespace AplicativoWebAcademia.Controllers
             {
                 return NotFound();
             }
-
             var pessoaModel = await _context.PessoaModel
                 .FirstOrDefaultAsync(m => m.Codigo == id);
             if (pessoaModel == null)
@@ -134,9 +178,15 @@ namespace AplicativoWebAcademia.Controllers
         // POST: PessoaModels/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int id) 
         {
+            
             var pessoaModel = await _context.PessoaModel.FindAsync(id);
+            if (PessoaRegras.VerificaAtivo(pessoaModel.Alterar))
+            {
+                ModelState.AddModelError("regras", "Não é possivel excluir uma pessoa na situação 'Ativo'");
+                return View(pessoaModel);
+            }
             _context.PessoaModel.Remove(pessoaModel);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -145,5 +195,24 @@ namespace AplicativoWebAcademia.Controllers
         {
             return _context.PessoaModel.Any(e => e.Codigo == id);
         }
+
+        // GET: PessoaModels/VerificaAlteração/5
+        [HttpGet]
+        public async Task<IActionResult> VerificaAlteração(int id)
+        {
+            var pessoaModel = await _context.PessoaModel.FindAsync(id);
+            if (pessoaModel.Alterar.Equals("Ativo"))
+            {
+                pessoaModel.Alterar = "Ativo";
+            }
+            else
+            {
+                pessoaModel.Alterar = "Inativo";
+            }
+            _context.Update(pessoaModel);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
     }
 }
